+++ routes/categories.ts
@@ -24,9 +24,9 @@ export default function(loader: EventLoader) {
       const result = event.categories.map((category) => {
         return {
           name: category.name,
-          distance: (category as any).distance,
-          ascent: (category as any).ascent,
-          controls: (category as any).controls,
+          distance: category.distance,
+          ascent: category.ascent,
+          controls: category.controls,
           runners: category.runners.length
         };
       });
