+++ routes/course.ts
@@ -14,18 +14,9 @@
  * limitations under the License.
  */
 import { Request, Response } from "express";
-import { ranking, parseTime } from "@rasifix/orienteering-utils";
-import { EventLoader, Category } from "../types/index.ts";
-import { Runner } from "@rasifix/orienteering-utils/lib/model/runner";
-
-interface Course {
-  id: string;
-  name: string;
-  distance: number;
-  ascent: number;
-  controls: number;
-  runners: Runner[];
-}
+import { ranking } from "@rasifix/orienteering-utils";
+import { EventLoader } from "../types/index.ts";
+import { buildCourseDetails } from "../services/course-builder.ts";
 
 export default function (loader: EventLoader) {
   return (req: Request, res: Response) => {
@@ -35,7 +26,7 @@ export default function (loader: EventLoader) {
     loader(
       id,
       (event) => {
-        const courses = defineCourses(event.categories);
+        const courses = buildCourseDetails(event.categories);
         const course = courses.find((course) => {
           return course.id === courseId;
         });
@@ -67,87 +58,4 @@ export default function (loader: EventLoader) {
       }
     );
   };
-}
-
-function defineCourses(categories: Category[]) {
-  const groupedCategories: { [key: string]: Category[] } = {};
-  categories.forEach((category) => {
-    if (category.runners.length === 0) {
-      // category without runners are ignored
-      return;
-    }
-
-    // find categories with identical courses
-    const controls = category.runners[0]
-      .splits!.map((split) => split.code)
-      .join("-");
-    if (!groupedCategories[controls]) {
-      groupedCategories[controls] = [];
-    }
-    groupedCategories[controls].push(category);
-  });
-
-  // build courses
-  const courses: Course[] = [];
-  Object.keys(groupedCategories).forEach((grouped) => {
-    const cats = groupedCategories[grouped];
-    let idx = 0;
-    const id = cats
-      .map((cat) => cat.name)
-      .sort()
-      .join("-");
-    courses.push({
-      id: id,
-      name: id,
-      distance: (cats[0] as any).distance,
-      ascent: (cats[0] as any).ascent,
-      controls: (cats[0] as any).controls,
-      runners: cats.reduce((prev, cat) => {
-        return prev.concat(
-          cat.runners.map((runner) => {
-            return {
-              id: "" + ++idx,
-              startTime: runner.starttime,
-              yearOfBirth: runner.yearOfBirth,
-              time: runner.time,
-              splits: runner.splits,
-              club: runner.club,
-              fullName: runner.fullName,
-              city: runner.city,
-              category: cat.name,
-            };
-          })
-        );
-      }, [] as any[]),
-    });
-  });
-
-  courses
-    .sort((c1, c2) => {
-      if (c1.id < c2.id) {
-        return -1;
-      } else if (c1.id > c2.id) {
-        return 1;
-      } else {
-        return 0;
-      }
-    })
-    .forEach((course) => {
-      // sort the runners according to their run time
-      course.runners.sort((r1, r2) => {
-        const t1 = parseTime(r1.time);
-        const t2 = parseTime(r2.time);
-        if (t1 === null && t2 === null) {
-          return 0;
-        } else if (t1 !== null && t2 === null) {
-          return -1;
-        } else if (t1 === null && t2 !== null) {
-          return 1;
-        } else {
-          return (parseTime(r1.time) ?? 0) - (parseTime(r2.time) ?? 0);
-        }
-      });
-    });
-
-  return courses;
 }
